reverted buzz to index
